Rework Algorithm of Thought strategy
- Full rework, possibly still not 100% functionally correct.

Thoughts: I noticed Kyes implementations seem to struggle a lot and that it is pretty far off the paper it is based on, so a full rework was required. I plan to remove my Kye ToT implementation from the project and attempt to abstract the official ToT implementation into the framework.